requirements.txt
irc
discord.py
pony
python-dotenv
pymysql
